(wip) start breaking up evaluation lib into smalller files
